refactoring into module
