+++ setup/eatup_setup_server/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "eatup_setup_server"
-version = "0.1.5"
+version = "0.1.7"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
